2019 - upgrade intcode to bigint

+++ 2019/lib/intcode/operations.c
@@ -1,11 +1,11 @@
 #include <stdio.h>
 #include <stdlib.h>
-#include "stdbool.h"
+#include <stdbool.h>
 #include "intcode.h"
 
 bool intcode_op_add(t_intcode_state* state) {
-    int lhs, rhs;
-    int op = state->memory.data[state->ip];
+    bigint lhs, rhs;
+    bigint op = state->memory.data[state->ip];
 
     return (state->ip + 3) < state->memory.size
            && intcode_safe_read(state, state->memory.data[state->ip + 1], intcode_get_mode(op, 1), &lhs)
@@ -14,8 +14,8 @@ bool intcode_op_add(t_intcode_state* state) {
 }
 
 bool intcode_op_mul(t_intcode_state* state) {
-    int lhs, rhs;
-    int op = state->memory.data[state->ip];
+    bigint lhs, rhs;
+    bigint op = state->memory.data[state->ip];
 
     return (state->ip + 3) < state->memory.size
            && intcode_safe_read(state, state->memory.data[state->ip + 1], intcode_get_mode(op, 1), &lhs)
@@ -24,8 +24,8 @@ bool intcode_op_mul(t_intcode_state* state) {
 }
 
 bool intcode_op_read(t_intcode_state* state) {
-    int input;
-    int op = state->memory.data[state->ip];
+    bigint input;
+    bigint op = state->memory.data[state->ip];
 
     return (state->ip + 1) < state->memory.size
            && intcode_safe_read_input(state, &input)
@@ -33,8 +33,8 @@ bool intcode_op_read(t_intcode_state* state) {
 }
 
 bool intcode_op_write(t_intcode_state* state) {
-    int output;
-    int op = state->memory.data[state->ip];
+    bigint output;
+    bigint op = state->memory.data[state->ip];
 
     return (state->ip + 1) < state->memory.size
            && intcode_safe_read(state, state->memory.data[state->ip + 1], intcode_get_mode(op, 1), &output)
@@ -42,8 +42,8 @@ bool intcode_op_write(t_intcode_state* state) {
 }
 
 bool _intcode_jump(t_intcode_state* state, bool expected) {
-    int test, val;
-    int op = state->memory.data[state->ip];
+    bigint test, val;
+    bigint op = state->memory.data[state->ip];
 
     if ((state->ip + 2) < state->memory.size
         && intcode_safe_read(state, state->memory.data[state->ip + 1], intcode_get_mode(op, 1), &test)
@@ -70,7 +70,7 @@ enum {
     INTCODE_CMP_EQ
 };
 
-int _intcode_cmp_operation(int operation, int lhs, int rhs) {
+int _intcode_cmp_operation(int operation, bigint lhs, bigint rhs) {
     switch (operation) {
         case INTCODE_CMP_LT:
             return lhs < rhs;
@@ -82,8 +82,8 @@ int _intcode_cmp_operation(int operation, int lhs, int rhs) {
 }
 
 bool _intcode_cmp(t_intcode_state* state, int operation) {
-    int lhs, rhs;
-    int op = state->memory.data[state->ip];
+    bigint lhs, rhs;
+    bigint op = state->memory.data[state->ip];
 
     return (state->ip + 3) < state->memory.size
            && intcode_safe_read(state, state->memory.data[state->ip + 1], intcode_get_mode(op, 1), &lhs)
@@ -100,8 +100,8 @@ bool intcode_op_eq(t_intcode_state* state) {
 }
 
 bool intcode_op_adjust_relative_base(t_intcode_state* state) {
-    int offset;
-    int op = state->memory.data[state->ip];
+    bigint offset;
+    bigint op = state->memory.data[state->ip];
 
     if ((state->ip + 1) < state->memory.size
         && intcode_safe_read(state, state->memory.data[state->ip + 1], intcode_get_mode(op, 1), &offset)) {
